Comments in the main.m and some embellishing to show results -Modify main.m file with some comments when the functions are called. The function procrustesAnalysis2() and principalComponentAnalysis() were commented
-Modify principalComponentAnalysis() to generate the approximation of the
sample and it is showed the approximation and the sample.

-Modify initializeDataSetHands() function to compy phi = shapes; instead of
just a matrix with zeros.

Change in how the data is loaded and processed
-Add initializeDataSetHands2() function to load the hands dataset as
just one matrix where the rows are the samples and the columns are the
points. Each sample has 112 points, the first 56 columns belons to x
and the nest 56 colums belongs to y.

-Add procrustesAnalysis2() function to perform Procrustes, but using the
dataset loaded with the initializeDataSetHands2() function.

Using these new versions, it is necessary to modify the
principalComponentAnalysis() function according to how the dataset has
been represented.