Initial commit for version drone-1.1.0
The main.m file was renamed as main_ssm.m and it contains the basic
methods to generate the SSM of the hands. The file main_ssm_cpd.m
was added to use it to generate the SSM with the Coherent Point Drift
(CPD) method. The files of this method are under the directory 'extra'.