Applying CPD and computing generalization
The function principalComponentAnalysis() returned some plots related
to the calculation of the approximation of one example of the dataset,
however it is not necessary for this function to do this. That is why
was modified to return just the eigenvalues and eigenvectors.

The function procrustesAnalysis() did not accept an initial mean shape,
this was necesary in order to use the CPD algorithm that returns a mean
shape and the registration of the shapes. This parameters are used to
initialization in procrustesAnalysis().

The files generalization*.m are used to probe how good is the SSM to
generalize. This files use the function leaveOneOutPartition(). This
files still need to to tested.